created profile component, moved CSS files to styles folder, updated imports for each component to reflect the css move
