+++ client/components/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Footer.css';
+import './styles/Footer.css';
 import { Button } from './Button';
 import { Link } from 'react-router-dom';
