+++ client/components/modal.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import './modal.css';
+import './styles/modal.css';
 
 const Modal = ({display, handleClose}) => {
