+++ client/components/profile.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import './styles/profile.css';
+
+const Profile = () => {
+
+return(
+  <div>
+    <section>
+
+    </section>
+  </div>
+)
+};
+export default Profile;
